refactor: move version parsing tests to parser
