+++ prql-compiler/src/parser/mod.rs
@@ -2743,4 +2743,43 @@ join s=salaries (==id)
         )
         .unwrap();
     }
+
+    #[test]
+    fn check_valid_version() {
+        let stmt = format!(
+            r#"
+        prql version:"{}"
+        "#,
+            env!("CARGO_PKG_VERSION_MAJOR")
+        );
+        assert!(parse_single(&stmt).is_ok());
+
+        let stmt = format!(
+            r#"
+            prql version:"{}.{}"
+            "#,
+            env!("CARGO_PKG_VERSION_MAJOR"),
+            env!("CARGO_PKG_VERSION_MINOR")
+        );
+        assert!(parse_single(&stmt).is_ok());
+
+        let stmt = format!(
+            r#"
+            prql version:"{}.{}.{}"
+            "#,
+            env!("CARGO_PKG_VERSION_MAJOR"),
+            env!("CARGO_PKG_VERSION_MINOR"),
+            env!("CARGO_PKG_VERSION_PATCH"),
+        );
+        assert!(parse_single(&stmt).is_ok());
+    }
+
+    #[test]
+    fn check_invalid_version() {
+        let stmt = format!(
+            "prql version:{}\n",
+            env!("CARGO_PKG_VERSION_MAJOR").parse::<usize>().unwrap() + 1
+        );
+        assert!(parse_single(&stmt).is_err());
+    }
 }
