+++ prql-compiler/src/semantic/mod.rs
@@ -164,7 +164,7 @@ pub mod test {
     use insta::assert_yaml_snapshot;
 
     use crate::ast::rq::Query;
-    use crate::parser::{parse, parse_single};
+    use crate::parser::parse;
 
     use super::{resolve, resolve_and_lower, Context};
 
@@ -295,44 +295,5 @@ pub mod test {
         "###,
         )
         .is_err());
-    }
-
-    #[test]
-    fn check_valid_version() {
-        let stmt = format!(
-            r#"
-        prql version:"{}"
-        "#,
-            env!("CARGO_PKG_VERSION_MAJOR")
-        );
-        assert!(parse_single(&stmt).is_ok());
-
-        let stmt = format!(
-            r#"
-            prql version:"{}.{}"
-            "#,
-            env!("CARGO_PKG_VERSION_MAJOR"),
-            env!("CARGO_PKG_VERSION_MINOR")
-        );
-        assert!(parse_single(&stmt).is_ok());
-
-        let stmt = format!(
-            r#"
-            prql version:"{}.{}.{}"
-            "#,
-            env!("CARGO_PKG_VERSION_MAJOR"),
-            env!("CARGO_PKG_VERSION_MINOR"),
-            env!("CARGO_PKG_VERSION_PATCH"),
-        );
-        assert!(parse_single(&stmt).is_ok());
-    }
-
-    #[test]
-    fn check_invalid_version() {
-        let stmt = format!(
-            "prql version:{}\n",
-            env!("CARGO_PKG_VERSION_MAJOR").parse::<usize>().unwrap() + 1
-        );
-        assert!(parse_single(&stmt).is_err());
     }
 }
